Remove the X-XSS-Protection rule
From https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-XSS-Protection:

Non-standard: This feature is non-standard and is not on a standards track. Do not use it on production sites facing the Web: it will not work for every user. There may also be large incompatibilities between implementations and the behavior may change in the future.